refactor(registry): extract shared load_registry into registry.py
Both topology_mapper.py and registry_audit.py had their own
load_registry implementations that had diverged (different error
handling, different path resolution). Extract into a shared module
with the improved error-string-return pattern and __file__-relative
DEFAULT_REGISTRY_PATH.
